test: Tidying up the tests (#3334)
(I'm struggling with the PR mechanism here as there the "remove datrie"
commits and changes are showing up, even though they should be made on
both branches. Anyway, I'm hoping I can open the PR first then clean it
up.)

The purpose of this change is to try and improve the tests suite by:

1) Providing `tests/README.md` to guide new developers how to run the
tests
1) Split out the tests that require conda to run into a separate
`tests_using_conda.py`
1) Allow the main tests to pass even if Snakemake cannot determine its
own version (and thinks it is version 0.0)

Also:

1) Removed dependency on stress by using stress-ng which is available
via Bioconda
1) Clean up some test files that should never have been committed to GIT
1) Change some 10 second stress delays to 4 seconds as there seems no
reason for them to be that long
1) Fix a test that modified the environment (`shell.prefix`) so later
tests could fail depending on the order of tests run
