+++ .gitignore
@@ -14,6 +14,7 @@ dist/
 .venv
 .venv/*
 .python-version
+/version.txt
 
 
 .idea
