+++ test-environment.yml
@@ -51,6 +51,7 @@ dependencies:
   - reretry
   - graphviz
   - humanfriendly
+  - stress-ng
   - throttler
   - dpath
   - nodejs # for cwltool
