+++ tests/README.md
@@ -1,0 +1,10 @@
+# Unit and regression tests for Snakemake
+
+All tests are written to by run by [pytest](https://docs.pytest.org/en/stable/). Note that not
+all tests in this directory are expected to pass, but the specific tests set up for continuous
+integration in `.github/workflows/main.yml` should all work.
+
+For info on setting up a working conda environment to run these tests, and other general guidance,
+see [the Snakemake contributing guide](
+https://snakemake.readthedocs.io/en/stable/project_info/contributing.html#testing-guidelines).
+
