--- tests/test_prebuilt_conda_script/dummy_package/src/dummy.egg-info/PKG-INFO
@@ -1,7 +1,0 @@
-Metadata-Version: 2.1
-Name: dummy
-Version: 0.0.0
-Author: Egor Kosaretskiy
-Classifier: Programming Language :: Python :: 3
-Classifier: License :: OSI Approved :: MIT License
-Classifier: Operating System :: OS Independent
