--- tests/test_prebuilt_conda_script/dummy_package/src/dummy.egg-info/SOURCES.txt
@@ -1,8 +1,0 @@
-pyproject.toml
-setup.cfg
-src/dummy/__init__.py
-src/dummy/dummy.py
-src/dummy.egg-info/PKG-INFO
-src/dummy.egg-info/SOURCES.txt
-src/dummy.egg-info/dependency_links.txt
-src/dummy.egg-info/top_level.txt
