--- tests/test_prebuilt_conda_script/dummy_package/src/dummy.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
