--- tests/test_prebuilt_conda_script/dummy_package/src/dummy.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-dummy
